requirements.txt
fastapi
uvicorn
pyserial
pygame
pytest
